v0.1.0: screen issue resolved
