+++ src/main/java/org/capital/LoginUser.java
@@ -12,28 +12,52 @@ public class LoginUser extends JFrame implements ActionListener {
     public LoginUser() {
         setTitle("Login");
         setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        setSize(300, 150);
+        setSize(500, 250);
         setLocationRelativeTo(null);
 
-        JLabel usernameLabel = new JLabel("Username:");
-        JLabel passwordLabel = new JLabel("Password:");
-        usernameField = new JTextField(20);
-        passwordField = new JPasswordField(20);
-        JButton loginButton = new JButton("Login");
+        JetBrainsMonoFont monoFont = new JetBrainsMonoFont();
+        monoFont.useJetBrainsMonoFont();
 
-        setLayout(new GridLayout(3, 2));
-        add(usernameLabel);
-        add(usernameField);
-        add(passwordLabel);
-        add(passwordField);
-        add(new JLabel());
-        add(loginButton);
+        Font jetBrainsMono = new Font("JetBrains Mono", Font.PLAIN, 12);
+
+        JLabel usernameLabel = new JLabel("Username:");
+        usernameLabel.setFont(jetBrainsMono);
+        JLabel passwordLabel = new JLabel("Password:");
+        passwordLabel.setFont(jetBrainsMono);
+        usernameField = new JTextField(20);
+        usernameField.setFont(jetBrainsMono);
+        passwordField = new JPasswordField(20);
+        passwordField.setFont(jetBrainsMono);
+        JButton loginButton = new JButton("Login");
+        loginButton.setFont(jetBrainsMono);
+
+        setLayout(new GridBagLayout());
+        GridBagConstraints gbc = new GridBagConstraints();
+        gbc.insets = new Insets(5, 5, 5, 5);
+
+        gbc.gridx = 0;
+        gbc.gridy = 0;
+        add(usernameLabel, gbc);
+
+        gbc.gridx = 1;
+        gbc.gridy = 0;
+        add(usernameField, gbc);
+
+        gbc.gridx = 0;
+        gbc.gridy = 1;
+        add(passwordLabel, gbc);
+
+        gbc.gridx = 1;
+        gbc.gridy = 1;
+        add(passwordField, gbc);
+
+        gbc.gridx = 0;
+        gbc.gridy = 2;
+        gbc.gridwidth = 2;
+        gbc.anchor = GridBagConstraints.CENTER;
+        add(loginButton, gbc);
 
         loginButton.addActionListener(this);
-    }
-
-    public static void main(String[] args) {
-        SwingUtilities.invokeLater(() -> new LoginUser().setVisible(true));
     }
 
     @Override
@@ -42,14 +66,16 @@ public class LoginUser extends JFrame implements ActionListener {
         String password = new String(passwordField.getPassword());
 
         if (isValidUser(username, password)) {
-            new Main().setVisible(true);
-            dispose();
+            this.setVisible(false);
+            Capital capital = new Capital();
+            capital.setVisible(true);
+
         } else {
             JOptionPane.showMessageDialog(this, "Invalid username or password", "Error", JOptionPane.ERROR_MESSAGE);
         }
     }
 
     private boolean isValidUser(String username, String password) {
-        return username.equals("vnt") && password.equals("seven");
+        return username.equals("capital") && password.equals("seven");
     }
 }
